cubecl_linalg::matmul

Module components

Components for matrix multiplication

Modules

as_cmma_layout

batch

global

stage

tile

Structs

FormattedConfigError

LhsStageDim

Dimensions for lhs stage.

LhsStageDimExpand

MatmulProblem

Description of a matmul problem to solve, regardless of actual data

MatmulSelection

MatmulSize

MatmulSizeExpand

OutStageDim

Dimensions for out stage.

OutStageDimExpand

RhsStageDim

Dimensions for rhs stage.

RhsStageDimExpand

SingleMatmulSpec

Specification for a simple standard matmul using global tensor as inputs.

StageDims

Aggregation of StageDims for all stages

Enums

Ident

Identifier for all three tensors in a matmul

InputIdent

MatrixLayout

Layout of a 2D structure such as a tensor, shared memory or slice, used within any matmul kernel level

MatrixLayoutExpand

Traits

MatmulConfig

A config for a matmul

MatmulConfigFactory

Provides configuration for a matmul kernel at any level

MatmulLaunch

Provides launch entry point to solve a matmul

MatmulPrecision

Matrix multiplication precisions.

MatmulSpec

Matrix multiplication spec definiting each element types used in the computation as well as how the arguments are passed to the kernel.

StageDim

Functions

as_cmma_layout

Maps the matmul MatrixLayout to cmma’s MatrixLayout, for use in Cmma API.

Type Aliases

InputArg

Input argument

InputRuntimeArg

Input runtime argument

InvalidConfigError

OutputArg

Output argument

OutputRuntimeArg

Output runtime argument