cubecl_linalg::matmul::components

Module batch

Modules

one_to_many

one_to_one

swizzle

Structs

ColMajorSpanMatmul

Iterates on global matmuls in a col major fashion

ColMajorSpanMatmulExpand

NaturalDispatch

Operates on data further along the m dimension as cube_pos_x increases, and further along the n dimension as cube_pos_y increases.

RowMajorSpanMatmul

Iterates on global matmuls in a row major fashion

RowMajorSpanMatmulExpand

Span

Area of a tensor a cube is responsible of performing matmul Similar to the concept of tensor slice, but specialized for matmul constraints

SpanDim

Span information in one dimension

SpanDimExpand

SpanExpand

SwizzleNaturalDispatch

Processes data in a swizzled pattern, prioritizing cubes along the x-axis first.

SwizzleSpanMatmul

Iterates on global matmuls following the swizzle algorithm

SwizzleSpanMatmulExpand

SwizzleTransposedDispatch

Processes data in a swizzled pattern, prioritizing cubes along the y-axis first.

TransposedDispatch

Operates on data further along the m dimension as cube_pos_x increases, and further along the n dimension as cube_pos_y increases.

Traits

BatchConfig

Configuration for the batch matmul level.

BatchMatmul

Provides matrix multiplication operations at the batch level.

BatchMatmulFamily

A family of matmuls working with any precision.

CubeCountDispatch

CubeDispatch

Distributes cube instances across the tensor, assigning each to compute data in distinct regions.

SpanMatmul

Iterates on several global matmul across a span

Functions

swizzle