cubecl_linalg::matmul::components::batch

Trait BatchConfig

pub trait BatchConfig: MatmulConfig {
    type GmmConfig: GlobalConfig;

    // Required methods
    fn to_gmm_config(&self) -> Self::GmmConfig;
    fn stage_dim(&self, ident: Ident) -> Box<dyn StageDim>;
    fn max_m(&self) -> u32;
    fn max_n(&self) -> u32;
    fn max_batches(&self) -> u32;
}

Configuration for the batch matmul level.

Required Associated Types

type GmmConfig: GlobalConfig

Underlying Global matmul config

Required Methods

fn to_gmm_config(&self) -> Self::GmmConfig

Convert itself to the underlying global matmul config

fn stage_dim(&self, ident: Ident) -> Box<dyn StageDim>

Returns the StageDim for the given ident

fn max_m(&self) -> u32

Returns the largest m dimension supported with these configs

fn max_n(&self) -> u32

Returns the largest n dimension supported with these configs

fn max_batches(&self) -> u32

Returns the largest number of batches supported with these configs

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.