cubecl_linalg::matmul::components

Module global

Modules

args

buffered

full_load

output_loader

tensor_view

Structs

CommonGlobalConfig

Configuration for the pipelined global matmul

CommonGlobalConfigExpand

ZeroAccumulatorLoader

Accumulator loader that zeros the accumulator

ZeroAccumulatorLoaderExpand

Traits

AccumulatorLoader

Input to the global matmul accumulator, responsible of filling the stage and providing a reader for it.

GlobalConfig

Configuration for the global matmul level.

GlobalMatmul

Provides matrix multiplication operations at the global level.

GlobalMatmulFamily

A family of matmuls working with any precision.

InputLoader

Input to the global matmul, responsible of filling the stage and providing a reader for it. Advances along the k-dimension to fill the stage with further data.

LoadingValidation

OutputLoader

Output to the global matmul