cubecl_linalg::matmul::components::global::args

Trait MatmulArgs

pub trait MatmulArgs:
    Send
    + Sync
    + 'static
    + Clone {
    type Input<EG: Numeric>: LaunchArg + CubeType;
    type Output<EG: Numeric>: LaunchArg + CubeType;
    type State<EG: Numeric>: CubeType;

    // Required methods
    fn init_state<EG: Numeric>(
        input: &Self::Input<EG>,
        output: &mut Self::Output<EG>,
    ) -> Self::State<EG>;
    fn read_lhs<EG: Numeric>(
        state: &Self::State<EG>,
        coordinate: u32,
    ) -> Line<EG>;
    fn read_rhs<EG: Numeric>(
        state: &Self::State<EG>,
        coordinate: u32,
    ) -> Line<EG>;
    fn write_out<EG: Numeric>(
        state: &mut Self::State<EG>,
        coordinate: u32,
        value: Line<EG>,
    );
    fn rank_lhs<EG: Numeric>(state: &Self::State<EG>) -> u32;
    fn rank_rhs<EG: Numeric>(state: &Self::State<EG>) -> u32;
    fn rank_out<EG: Numeric>(state: &Self::State<EG>) -> u32;
    fn shape_lhs<EG: Numeric>(state: &Self::State<EG>, axis: u32) -> u32;
    fn shape_rhs<EG: Numeric>(state: &Self::State<EG>, axis: u32) -> u32;
    fn shape_out<EG: Numeric>(state: &Self::State<EG>, axis: u32) -> u32;
    fn stride_lhs<EG: Numeric>(state: &Self::State<EG>, axis: u32) -> u32;
    fn stride_rhs<EG: Numeric>(state: &Self::State<EG>, axis: u32) -> u32;
    fn stride_out<EG: Numeric>(state: &Self::State<EG>, axis: u32) -> u32;
    fn __expand_init_state<EG: Numeric>(
        context: &mut CubeContext,
        input: <Self::Input<EG> as CubeType>::ExpandType,
        output: <Self::Output<EG> as CubeType>::ExpandType,
    ) -> <Self::State<EG> as CubeType>::ExpandType;
    fn __expand_read_lhs<EG: Numeric>(
        context: &mut CubeContext,
        state: <Self::State<EG> as CubeType>::ExpandType,
        coordinate: <u32 as CubeType>::ExpandType,
    ) -> <Line<EG> as CubeType>::ExpandType;
    fn __expand_read_rhs<EG: Numeric>(
        context: &mut CubeContext,
        state: <Self::State<EG> as CubeType>::ExpandType,
        coordinate: <u32 as CubeType>::ExpandType,
    ) -> <Line<EG> as CubeType>::ExpandType;
    fn __expand_write_out<EG: Numeric>(
        context: &mut CubeContext,
        state: <Self::State<EG> as CubeType>::ExpandType,
        coordinate: <u32 as CubeType>::ExpandType,
        value: <Line<EG> as CubeType>::ExpandType,
    ) -> <() as CubeType>::ExpandType;
    fn __expand_rank_lhs<EG: Numeric>(
        context: &mut CubeContext,
        state: <Self::State<EG> as CubeType>::ExpandType,
    ) -> <u32 as CubeType>::ExpandType;
    fn __expand_rank_rhs<EG: Numeric>(
        context: &mut CubeContext,
        state: <Self::State<EG> as CubeType>::ExpandType,
    ) -> <u32 as CubeType>::ExpandType;
    fn __expand_rank_out<EG: Numeric>(
        context: &mut CubeContext,
        state: <Self::State<EG> as CubeType>::ExpandType,
    ) -> <u32 as CubeType>::ExpandType;
    fn __expand_shape_lhs<EG: Numeric>(
        context: &mut CubeContext,
        state: <Self::State<EG> as CubeType>::ExpandType,
        axis: <u32 as CubeType>::ExpandType,
    ) -> <u32 as CubeType>::ExpandType;
    fn __expand_shape_rhs<EG: Numeric>(
        context: &mut CubeContext,
        state: <Self::State<EG> as CubeType>::ExpandType,
        axis: <u32 as CubeType>::ExpandType,
    ) -> <u32 as CubeType>::ExpandType;
    fn __expand_shape_out<EG: Numeric>(
        context: &mut CubeContext,
        state: <Self::State<EG> as CubeType>::ExpandType,
        axis: <u32 as CubeType>::ExpandType,
    ) -> <u32 as CubeType>::ExpandType;
    fn __expand_stride_lhs<EG: Numeric>(
        context: &mut CubeContext,
        state: <Self::State<EG> as CubeType>::ExpandType,
        axis: <u32 as CubeType>::ExpandType,
    ) -> <u32 as CubeType>::ExpandType;
    fn __expand_stride_rhs<EG: Numeric>(
        context: &mut CubeContext,
        state: <Self::State<EG> as CubeType>::ExpandType,
        axis: <u32 as CubeType>::ExpandType,
    ) -> <u32 as CubeType>::ExpandType;
    fn __expand_stride_out<EG: Numeric>(
        context: &mut CubeContext,
        state: <Self::State<EG> as CubeType>::ExpandType,
        axis: <u32 as CubeType>::ExpandType,
    ) -> <u32 as CubeType>::ExpandType;

}

Arguments for the matrix multiplication algorithm.

Required Associated Types

type Input<EG: Numeric>: LaunchArg + CubeType

Type used for the input.

type Output<EG: Numeric>: LaunchArg + CubeType

Type used for the output.

type State<EG: Numeric>: CubeType

Inner state that is used to create tensor inputs and tensor outputs .

Required Methods

fn init_state<EG: Numeric>( input: &Self::Input<EG>, output: &mut Self::Output<EG>, ) -> Self::State<EG>

Init the state.

fn read_lhs<EG: Numeric>(state: &Self::State<EG>, coordinate: u32) -> Line<EG>

Read the line of the lhs tensor using the state at the given coordinate.

fn read_rhs<EG: Numeric>(state: &Self::State<EG>, coordinate: u32) -> Line<EG>

Read the line of the rhs tensor using the state at the given coordinate.

fn write_out<EG: Numeric>( state: &mut Self::State<EG>, coordinate: u32, value: Line<EG>, )

Write the line to the output at the given coordinate using the state.

fn rank_lhs<EG: Numeric>(state: &Self::State<EG>) -> u32

Get the rank of the lhs tensor using the state.

fn rank_rhs<EG: Numeric>(state: &Self::State<EG>) -> u32

Get the rank of the rhs tensor using the state.

fn rank_out<EG: Numeric>(state: &Self::State<EG>) -> u32

Get the rank of the out tensor using the state.

fn shape_lhs<EG: Numeric>(state: &Self::State<EG>, axis: u32) -> u32

Get the shape of the lhs tensor using the state.

fn shape_rhs<EG: Numeric>(state: &Self::State<EG>, axis: u32) -> u32

Get the shape of the rhs tensor using the state.

fn shape_out<EG: Numeric>(state: &Self::State<EG>, axis: u32) -> u32

Get the shape of the out tensor using the state.

fn stride_lhs<EG: Numeric>(state: &Self::State<EG>, axis: u32) -> u32

Get the stride of the lhs tensor using the state.

fn stride_rhs<EG: Numeric>(state: &Self::State<EG>, axis: u32) -> u32

Get the stride of the rhs tensor using the state.

fn stride_out<EG: Numeric>(state: &Self::State<EG>, axis: u32) -> u32

Get the stride of the out tensor using the state.

fn __expand_init_state<EG: Numeric>( context: &mut CubeContext, input: <Self::Input<EG> as CubeType>::ExpandType, output: <Self::Output<EG> as CubeType>::ExpandType, ) -> <Self::State<EG> as CubeType>::ExpandType

fn __expand_read_lhs<EG: Numeric>( context: &mut CubeContext, state: <Self::State<EG> as CubeType>::ExpandType, coordinate: <u32 as CubeType>::ExpandType, ) -> <Line<EG> as CubeType>::ExpandType

fn __expand_read_rhs<EG: Numeric>( context: &mut CubeContext, state: <Self::State<EG> as CubeType>::ExpandType, coordinate: <u32 as CubeType>::ExpandType, ) -> <Line<EG> as CubeType>::ExpandType

fn __expand_write_out<EG: Numeric>( context: &mut CubeContext, state: <Self::State<EG> as CubeType>::ExpandType, coordinate: <u32 as CubeType>::ExpandType, value: <Line<EG> as CubeType>::ExpandType, ) -> <() as CubeType>::ExpandType

fn __expand_rank_lhs<EG: Numeric>( context: &mut CubeContext, state: <Self::State<EG> as CubeType>::ExpandType, ) -> <u32 as CubeType>::ExpandType

fn __expand_rank_rhs<EG: Numeric>( context: &mut CubeContext, state: <Self::State<EG> as CubeType>::ExpandType, ) -> <u32 as CubeType>::ExpandType

fn __expand_rank_out<EG: Numeric>( context: &mut CubeContext, state: <Self::State<EG> as CubeType>::ExpandType, ) -> <u32 as CubeType>::ExpandType

fn __expand_shape_lhs<EG: Numeric>( context: &mut CubeContext, state: <Self::State<EG> as CubeType>::ExpandType, axis: <u32 as CubeType>::ExpandType, ) -> <u32 as CubeType>::ExpandType

fn __expand_shape_rhs<EG: Numeric>( context: &mut CubeContext, state: <Self::State<EG> as CubeType>::ExpandType, axis: <u32 as CubeType>::ExpandType, ) -> <u32 as CubeType>::ExpandType

fn __expand_shape_out<EG: Numeric>( context: &mut CubeContext, state: <Self::State<EG> as CubeType>::ExpandType, axis: <u32 as CubeType>::ExpandType, ) -> <u32 as CubeType>::ExpandType

fn __expand_stride_lhs<EG: Numeric>( context: &mut CubeContext, state: <Self::State<EG> as CubeType>::ExpandType, axis: <u32 as CubeType>::ExpandType, ) -> <u32 as CubeType>::ExpandType

fn __expand_stride_rhs<EG: Numeric>( context: &mut CubeContext, state: <Self::State<EG> as CubeType>::ExpandType, axis: <u32 as CubeType>::ExpandType, ) -> <u32 as CubeType>::ExpandType

fn __expand_stride_out<EG: Numeric>( context: &mut CubeContext, state: <Self::State<EG> as CubeType>::ExpandType, axis: <u32 as CubeType>::ExpandType, ) -> <u32 as CubeType>::ExpandType

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.