cubecl_linalg::matmul::components::global::buffered::pipelined

Struct PipelinedMatmul

pub struct PipelinedMatmul<MP: MatmulPrecision, SMM: StageMatmul<MP::ES, MP::EG, MP::EA>> { /* private fields */ }

Performs matrix multiplication at the global level, with planes pipelining their work using two buffers: While they trigger a load event from global memory to shared memory on buffer A, they trigger a computation event from tensor cores on buffer B. Then buffers are switched.