cubecl_linalg::matmul::components::global::tensor_view

Struct TensorReader

pub struct TensorReader<E: Numeric> {
    pub tensor: VirtualTensor<E>,
    pub x_offset: u32,
    pub y_offset: u32,
    pub stride_x: u32,
    pub stride_y: u32,
    pub shape_x: u32,
    pub shape_y: u32,
    pub batch_offset: u32,
}

A view of a tensor that starts reading data from a specified offset. Ensures safe access by preventing out-of-bounds errors. Includes pre-fetched shapes and strides for optimized performance.

Fields
tensor: VirtualTensor<E>x_offset: u32y_offset: u32stride_x: u32stride_y: u32shape_x: u32shape_y: u32batch_offset: u32
Implementations

impl<EG: Numeric> TensorReader<EG>

pub fn new( tensor: VirtualTensor<EG>, x_offset: u32, y_offset: u32, batch_offset: u32, ) -> Self

Instantiate a read view over the given tensor, pre-fetching needed strides and shapes

pub fn update_view(&mut self, k_offset: u32, ident: Ident)

Advance the view along the k dimension by a specified offset, k_offset.

pub fn load_coalesced<G: GlobalConfig>( &self, tile_x: u32, tile_y: u32, unit_id: u32, ident: Ident, config: G, ) -> Line<EG>

Reads data from the tensor view at the specified tile coordinates (tile_x, tile_y).

Each unit loads one line in a coalesced manner for improved efficiency. For row-major tensors, subsequent units read lines horizontally within the tile, while for column-major tensors, they read lines vertically.

Note

Out-of-bounds reads will be translated to zeros.

pub fn __expand_new( context: &mut CubeContext, tensor: <VirtualTensor<EG> as CubeType>::ExpandType, x_offset: <u32 as CubeType>::ExpandType, y_offset: <u32 as CubeType>::ExpandType, batch_offset: <u32 as CubeType>::ExpandType, ) -> <Self as CubeType>::ExpandType

pub fn __expand_update_view( context: &mut CubeContext, this: <Self as CubeType>::ExpandType, k_offset: <u32 as CubeType>::ExpandType, ident: Ident, ) -> <() as CubeType>::ExpandType

pub fn __expand_load_coalesced<G: GlobalConfig>( context: &mut CubeContext, this: <Self as CubeType>::ExpandType, tile_x: <u32 as CubeType>::ExpandType, tile_y: <u32 as CubeType>::ExpandType, unit_id: <u32 as CubeType>::ExpandType, ident: Ident, config: G, ) -> <Line<EG> as CubeType>::ExpandType