cubecl_linalg::matmul::components::global

Trait AccumulatorLoader

pub trait AccumulatorLoader<O: Numeric, Acc: Numeric, G: StageConfig>:
    CubeType
    + 'static
    + Send
    + Sync {
    // Required methods
    fn fill_stage(this: &mut Self, config: G);
    fn load<I: Numeric, Tile: TileMatmul<I, Acc>>(
        this: &mut Self,
        acc: &mut Tile::Accumulator,
        tile_n: u32,
        config: Tile::Config,
    );
    fn __expand_fill_stage(
        context: &mut CubeContext,
        this: <Self as CubeType>::ExpandType,
        config: G,
    ) -> <() as CubeType>::ExpandType;
    fn __expand_load<I: Numeric, Tile: TileMatmul<I, Acc>>(
        context: &mut CubeContext,
        this: <Self as CubeType>::ExpandType,
        acc: <Tile::Accumulator as CubeType>::ExpandType,
        tile_n: <u32 as CubeType>::ExpandType,
        config: Tile::Config,
    ) -> <() as CubeType>::ExpandType;
}

Input to the global matmul accumulator, responsible of filling the stage and providing a reader for it.

Required Methods

fn fill_stage(this: &mut Self, config: G)

fn load<I: Numeric, Tile: TileMatmul<I, Acc>>( this: &mut Self, acc: &mut Tile::Accumulator, tile_n: u32, config: Tile::Config, )

Load accumulator for tile_n. Should call either zero_accumulator or fill_accumulator for the underlying tile.

fn __expand_fill_stage( context: &mut CubeContext, this: <Self as CubeType>::ExpandType, config: G, ) -> <() as CubeType>::ExpandType

fn __expand_load<I: Numeric, Tile: TileMatmul<I, Acc>>( context: &mut CubeContext, this: <Self as CubeType>::ExpandType, acc: <Tile::Accumulator as CubeType>::ExpandType, tile_n: <u32 as CubeType>::ExpandType, config: Tile::Config, ) -> <() as CubeType>::ExpandType

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.