cubecl_linalg::matmul::components::global

Trait GlobalConfig

pub trait GlobalConfig: MatmulConfig {
    type SmmConfig: StageConfig;

    // Required methods
    fn to_smm_config(&self) -> Self::SmmConfig;
    fn global_line_size(&self, ident: Ident) -> u32;
    fn stage_line_size(&self, ident: Ident) -> u32;
    fn stage_dim(&self, ident: Ident) -> Box<dyn StageDim>;
    fn layout(&self, ident: Ident) -> MatrixLayout;
    fn num_planes(&self) -> u32;
    fn plane_dim(&self) -> u32;
    fn tiling_order(&self, ident: Ident) -> TilingOrderConfig;
    fn check_m_bounds(&self) -> bool;
    fn check_k_bounds(&self) -> bool;
    fn check_n_bounds(&self) -> bool;
    fn transpose_load(&self, ident: Ident) -> bool;
}

Configuration for the global matmul level.

Required Associated Types

type SmmConfig: StageConfig

Underlying Stage matmul config

Required Methods

fn to_smm_config(&self) -> Self::SmmConfig

Convert itself to the underlying stage matmul config

fn global_line_size(&self, ident: Ident) -> u32

Returns the line size for the global memory corresponding to the given ident

fn stage_line_size(&self, ident: Ident) -> u32

Returns the line size for the stage of the given ident

fn stage_dim(&self, ident: Ident) -> Box<dyn StageDim>

Returns the StageDim for the given ident

fn layout(&self, ident: Ident) -> MatrixLayout

Returns the MatrixLayout for the given ident

fn num_planes(&self) -> u32

Returns the number of planes in the cube

fn plane_dim(&self) -> u32

Returns the size of the plane dimension

fn tiling_order(&self, ident: Ident) -> TilingOrderConfig

Returns the order in which tiles should be loaded to the stage

fn check_m_bounds(&self) -> bool

Whether it is necessary to add bound checks in the m dimension

fn check_k_bounds(&self) -> bool

Whether it is necessary to add bound checks in the k dimension

fn check_n_bounds(&self) -> bool

Whether it is necessary to add bound checks in the n dimension

fn transpose_load(&self, ident: Ident) -> bool

Whether we transpose data when loading to the stage

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.