cubecl_linalg::matmul::components::global

Trait GlobalMatmul

pub trait GlobalMatmul<MP: MatmulPrecision>:
    'static
    + Send
    + Sync {
    type Config: GlobalConfig;
    type LhsLoader: InputLoader<MP::EG, MP::ES, Self::Config>;
    type RhsLoader: InputLoader<MP::EG, MP::ES, Self::Config>;
    type AccumulatorLoader: CubeType;
    type Out: OutputLoader<MP::EG>;
    type Accumulator: CubeType;

    // Required methods
    fn execute(
        lhs_loader: Self::LhsLoader,
        rhs_loader: Self::RhsLoader,
        unloader: Self::Out,
        acc: &mut Self::Accumulator,
        k_range: (u32, u32),
        config: Self::Config,
    );
    fn init_lhs_loader(
        lhs: VirtualTensor<MP::EG>,
        m_offset: u32,
        k_offset: u32,
        batch_offset: u32,
        config: Self::Config,
    ) -> Self::LhsLoader;
    fn init_rhs_loader(
        rhs: VirtualTensor<MP::EG>,
        k_offset: u32,
        n_offset: u32,
        batch_offset: u32,
        config: Self::Config,
    ) -> Self::RhsLoader;
    fn init_unloader(
        out: VirtualTensor<MP::EG, ReadWrite>,
        m_offset: u32,
        n_offset: u32,
        batch_offset: u32,
    ) -> Self::Out;
    fn init_accumulator(config: Self::Config) -> Self::Accumulator;
    fn zero_accumulator(acc: &mut Self::Accumulator, config: Self::Config);
    fn __expand_execute(
        context: &mut CubeContext,
        lhs_loader: <Self::LhsLoader as CubeType>::ExpandType,
        rhs_loader: <Self::RhsLoader as CubeType>::ExpandType,
        unloader: <Self::Out as CubeType>::ExpandType,
        acc: <Self::Accumulator as CubeType>::ExpandType,
        k_range: <(u32, u32) as CubeType>::ExpandType,
        config: Self::Config,
    ) -> <() as CubeType>::ExpandType;
    fn __expand_init_lhs_loader(
        context: &mut CubeContext,
        lhs: <VirtualTensor<MP::EG> as CubeType>::ExpandType,
        m_offset: <u32 as CubeType>::ExpandType,
        k_offset: <u32 as CubeType>::ExpandType,
        batch_offset: <u32 as CubeType>::ExpandType,
        config: Self::Config,
    ) -> <Self::LhsLoader as CubeType>::ExpandType;
    fn __expand_init_rhs_loader(
        context: &mut CubeContext,
        rhs: <VirtualTensor<MP::EG> as CubeType>::ExpandType,
        k_offset: <u32 as CubeType>::ExpandType,
        n_offset: <u32 as CubeType>::ExpandType,
        batch_offset: <u32 as CubeType>::ExpandType,
        config: Self::Config,
    ) -> <Self::RhsLoader as CubeType>::ExpandType;
    fn __expand_init_unloader(
        context: &mut CubeContext,
        out: <VirtualTensor<MP::EG, ReadWrite> as CubeType>::ExpandType,
        m_offset: <u32 as CubeType>::ExpandType,
        n_offset: <u32 as CubeType>::ExpandType,
        batch_offset: <u32 as CubeType>::ExpandType,
    ) -> <Self::Out as CubeType>::ExpandType;
    fn __expand_init_accumulator(
        context: &mut CubeContext,
        config: Self::Config,
    ) -> <Self::Accumulator as CubeType>::ExpandType;
    fn __expand_zero_accumulator(
        context: &mut CubeContext,
        acc: <Self::Accumulator as CubeType>::ExpandType,
        config: Self::Config,
    ) -> <() as CubeType>::ExpandType;
}

Provides matrix multiplication operations at the global level.

At the global level,

Inputs are views over global memory, meaning access is given to only parts of the global memory inputs at once.

All planes within a Cube can collaborate to solve the problem

Dimensions M and N are fixed to an integer, but K is arbitrary large. The matrix multiplication works only for size (M, ) · (, N) = (M, N). M and N should match the underlying Stage matmul’s M and N.

Assumptions

Line sizes of the inputs evenly divide the dimension they are aligned with.

Safety

It is not assumed that the matmul’s dimensions match its inputs dimensions perfectly. It is therefore important that Loaders and Unloaders perform checks to avoid out-of-bounds before loading data.

Required Associated Types

type Config: GlobalConfig

type LhsLoader: InputLoader<MP::EG, MP::ES, Self::Config>

type RhsLoader: InputLoader<MP::EG, MP::ES, Self::Config>

type AccumulatorLoader: CubeType

type Out: OutputLoader<MP::EG>

type Accumulator: CubeType

Required Methods

fn execute( lhs_loader: Self::LhsLoader, rhs_loader: Self::RhsLoader, unloader: Self::Out, acc: &mut Self::Accumulator, k_range: (u32, u32), config: Self::Config, )

Performs the matrix multiplication over data loaded by the LHS and RHS loaders, over the range given for K, and stores with using the output unloader.

To compute the whole range of k values, use k_range=(0, K) where K is the K dimension of LHS and RHS.

fn init_lhs_loader( lhs: VirtualTensor<MP::EG>, m_offset: u32, k_offset: u32, batch_offset: u32, config: Self::Config, ) -> Self::LhsLoader

Initialize the loader for Lhs, starting at row m and column k

fn init_rhs_loader( rhs: VirtualTensor<MP::EG>, k_offset: u32, n_offset: u32, batch_offset: u32, config: Self::Config, ) -> Self::RhsLoader

Initialize the loader for Rhs, starting at row k and column n

fn init_unloader( out: VirtualTensor<MP::EG, ReadWrite>, m_offset: u32, n_offset: u32, batch_offset: u32, ) -> Self::Out

Initialize the unloader at row m and column n

fn init_accumulator(config: Self::Config) -> Self::Accumulator

Initialize the accumulator without data

fn zero_accumulator(acc: &mut Self::Accumulator, config: Self::Config)

Fill the accumulator with zeros

fn __expand_execute( context: &mut CubeContext, lhs_loader: <Self::LhsLoader as CubeType>::ExpandType, rhs_loader: <Self::RhsLoader as CubeType>::ExpandType, unloader: <Self::Out as CubeType>::ExpandType, acc: <Self::Accumulator as CubeType>::ExpandType, k_range: <(u32, u32) as CubeType>::ExpandType, config: Self::Config, ) -> <() as CubeType>::ExpandType

fn __expand_init_lhs_loader( context: &mut CubeContext, lhs: <VirtualTensor<MP::EG> as CubeType>::ExpandType, m_offset: <u32 as CubeType>::ExpandType, k_offset: <u32 as CubeType>::ExpandType, batch_offset: <u32 as CubeType>::ExpandType, config: Self::Config, ) -> <Self::LhsLoader as CubeType>::ExpandType

fn __expand_init_rhs_loader( context: &mut CubeContext, rhs: <VirtualTensor<MP::EG> as CubeType>::ExpandType, k_offset: <u32 as CubeType>::ExpandType, n_offset: <u32 as CubeType>::ExpandType, batch_offset: <u32 as CubeType>::ExpandType, config: Self::Config, ) -> <Self::RhsLoader as CubeType>::ExpandType

fn __expand_init_unloader( context: &mut CubeContext, out: <VirtualTensor<MP::EG, ReadWrite> as CubeType>::ExpandType, m_offset: <u32 as CubeType>::ExpandType, n_offset: <u32 as CubeType>::ExpandType, batch_offset: <u32 as CubeType>::ExpandType, ) -> <Self::Out as CubeType>::ExpandType

fn __expand_init_accumulator( context: &mut CubeContext, config: Self::Config, ) -> <Self::Accumulator as CubeType>::ExpandType

fn __expand_zero_accumulator( context: &mut CubeContext, acc: <Self::Accumulator as CubeType>::ExpandType, config: Self::Config, ) -> <() as CubeType>::ExpandType

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.