cubecl_linalg::matmul::components

Module stage

Modules

multi_buffer

single_buffer

Structs

ColMajorTiling

Tiles are conceptually stored in column-major order, regardless of the actual data layout.

CommonStageInput

RowMajorTiling

Tiles are conceptually stored in row-major order, regardless of the actual data layout.

Stage

Wrapper over the shared memory used for staging, abstracting its layout

Enums

TilingOrderConfig

Config determining which existing TilingOrder to use

TilingOrderConfigExpand

Traits

ReaderFamily

StageConfig

Configuration for the Stage matmul (SMM) level

StageMatmul

Provides matrix multiplication operations at the stage level.

StageMatmulFamily

StageReader

Input to the stage matmul, responsible of handing slices of data at precise locations in the stage

StageWriter

Responsible of writing the accumulated stage matmul output to global memory

TilingOrder

Layout in which to store tiles within the stage