cubecl_linalg::matmul::components::stage

Module single_buffer

Structs

LhsBufferReader

LhsBufferReaderExpand

LhsBufferReaderFamily

RhsBufferReader

RhsBufferReaderExpand

RhsBufferReaderFamily

SingleBufferMatmul

Performs matrix multiplication at the stage level, where each plane is responsible for a row of tiles:

SingleBufferMatmulFamily