cubecl_linalg::matmul::components::tile

Trait TileMatmul

pub trait TileMatmul<I: Numeric, O: Numeric>:
    'static
    + Send
    + Sync {
    type Config: TileConfig;
    type Lhs: CubeType;
    type Rhs: CubeType;
    type Accumulator: CubeType;

    // Required methods
    fn execute(
        lhs: &Self::Lhs,
        rhs: &Self::Rhs,
        out: &mut Self::Accumulator,
        config: Self::Config,
    );
    fn init_lhs(config: Self::Config) -> Self::Lhs;
    fn init_rhs(config: Self::Config) -> Self::Rhs;
    fn fill_lhs(
        slice: &Slice<Line<I>>,
        lhs: &mut Self::Lhs,
        config: Self::Config,
    );
    fn fill_rhs(
        slice: &Slice<Line<I>>,
        rhs: &mut Self::Rhs,
        config: Self::Config,
    );
    fn fill_accumulator(
        slice: &Slice<Line<O>>,
        acc: &mut Self::Accumulator,
        stride: u32,
        config: Self::Config,
    );
    fn read_accumulator<C: Numeric>(
        out: &Self::Accumulator,
        slice: &mut SliceMut<Line<C>>,
        config: Self::Config,
    );
    fn init_accumulator(config: Self::Config) -> Self::Accumulator;
    fn zero_accumulator(acc: &mut Self::Accumulator, config: Self::Config);
    fn __expand_execute(
        context: &mut CubeContext,
        lhs: <Self::Lhs as CubeType>::ExpandType,
        rhs: <Self::Rhs as CubeType>::ExpandType,
        out: <Self::Accumulator as CubeType>::ExpandType,
        config: Self::Config,
    ) -> <() as CubeType>::ExpandType;
    fn __expand_init_lhs(
        context: &mut CubeContext,
        config: Self::Config,
    ) -> <Self::Lhs as CubeType>::ExpandType;
    fn __expand_init_rhs(
        context: &mut CubeContext,
        config: Self::Config,
    ) -> <Self::Rhs as CubeType>::ExpandType;
    fn __expand_fill_lhs(
        context: &mut CubeContext,
        slice: <Slice<Line<I>> as CubeType>::ExpandType,
        lhs: <Self::Lhs as CubeType>::ExpandType,
        config: Self::Config,
    ) -> <() as CubeType>::ExpandType;
    fn __expand_fill_rhs(
        context: &mut CubeContext,
        slice: <Slice<Line<I>> as CubeType>::ExpandType,
        rhs: <Self::Rhs as CubeType>::ExpandType,
        config: Self::Config,
    ) -> <() as CubeType>::ExpandType;
    fn __expand_fill_accumulator(
        context: &mut CubeContext,
        slice: <Slice<Line<O>> as CubeType>::ExpandType,
        acc: <Self::Accumulator as CubeType>::ExpandType,
        stride: <u32 as CubeType>::ExpandType,
        config: Self::Config,
    ) -> <() as CubeType>::ExpandType;
    fn __expand_read_accumulator<C: Numeric>(
        context: &mut CubeContext,
        out: <Self::Accumulator as CubeType>::ExpandType,
        slice: <SliceMut<Line<C>> as CubeType>::ExpandType,
        config: Self::Config,
    ) -> <() as CubeType>::ExpandType;
    fn __expand_init_accumulator(
        context: &mut CubeContext,
        config: Self::Config,
    ) -> <Self::Accumulator as CubeType>::ExpandType;
    fn __expand_zero_accumulator(
        context: &mut CubeContext,
        acc: <Self::Accumulator as CubeType>::ExpandType,
        config: Self::Config,
    ) -> <() as CubeType>::ExpandType;

}

Provides matrix multiplication operations at the tile level.

At the tile level,

Inputs are raw slices of data, called tiles.

units within one plane can collaborate to solve the problem

dimensions M, N and K are fixed to an integer, and the matrix multiplication works only for size (M, K) · (K, N) = (M, N).

Assumptions:

Slices given as inputs must always be valid. If the actual matrix multiplication should be done on smaller sizes than M, N and K, padding with zeros must be done beforehand.

Enough units are present to perform the whole computation

Required Associated Types

type Config: TileConfig

type Lhs: CubeType

Contains LHS data that can be split across the units

type Rhs: CubeType

Contains RHS data that can be split across the units

type Accumulator: CubeType

Contains output data that can be split across the units

Required Methods

fn execute( lhs: &Self::Lhs, rhs: &Self::Rhs, out: &mut Self::Accumulator, config: Self::Config, )

Executes the matrix multiplication of LHS and RHS, adding the result to the output

fn init_lhs(config: Self::Config) -> Self::Lhs

Create the container for LHS data

Safety

This may point towards uninitialized memory. Make sure to call fill_lhs prior to execute.

fn init_rhs(config: Self::Config) -> Self::Rhs

Create the container for RHS data

Safety

This may point towards uninitialized memory. Make sure to call fill_rhs prior to execute.

fn fill_lhs(slice: &Slice<Line<I>>, lhs: &mut Self::Lhs, config: Self::Config)

Fill the container of LHS with data

fn fill_rhs(slice: &Slice<Line<I>>, rhs: &mut Self::Rhs, config: Self::Config)

Fill the container of RHS with data

fn fill_accumulator( slice: &Slice<Line<O>>, acc: &mut Self::Accumulator, stride: u32, config: Self::Config, )

Fill the accumulator with data

fn read_accumulator<C: Numeric>( out: &Self::Accumulator, slice: &mut SliceMut<Line<C>>, config: Self::Config, )

Write the content of the output container to the given slice

fn init_accumulator(config: Self::Config) -> Self::Accumulator

Create the container to receive the execution output.

Safety

The output container must be initialized to some value (typically 0), because the execution adds to the already present value.

fn zero_accumulator(acc: &mut Self::Accumulator, config: Self::Config)

Set the accumulator to zeros

fn __expand_execute( context: &mut CubeContext, lhs: <Self::Lhs as CubeType>::ExpandType, rhs: <Self::Rhs as CubeType>::ExpandType, out: <Self::Accumulator as CubeType>::ExpandType, config: Self::Config, ) -> <() as CubeType>::ExpandType

fn __expand_init_lhs( context: &mut CubeContext, config: Self::Config, ) -> <Self::Lhs as CubeType>::ExpandType

fn __expand_init_rhs( context: &mut CubeContext, config: Self::Config, ) -> <Self::Rhs as CubeType>::ExpandType

fn __expand_fill_lhs( context: &mut CubeContext, slice: <Slice<Line<I>> as CubeType>::ExpandType, lhs: <Self::Lhs as CubeType>::ExpandType, config: Self::Config, ) -> <() as CubeType>::ExpandType

fn __expand_fill_rhs( context: &mut CubeContext, slice: <Slice<Line<I>> as CubeType>::ExpandType, rhs: <Self::Rhs as CubeType>::ExpandType, config: Self::Config, ) -> <() as CubeType>::ExpandType

fn __expand_fill_accumulator( context: &mut CubeContext, slice: <Slice<Line<O>> as CubeType>::ExpandType, acc: <Self::Accumulator as CubeType>::ExpandType, stride: <u32 as CubeType>::ExpandType, config: Self::Config, ) -> <() as CubeType>::ExpandType

fn __expand_read_accumulator<C: Numeric>( context: &mut CubeContext, out: <Self::Accumulator as CubeType>::ExpandType, slice: <SliceMut<Line<C>> as CubeType>::ExpandType, config: Self::Config, ) -> <() as CubeType>::ExpandType

fn __expand_init_accumulator( context: &mut CubeContext, config: Self::Config, ) -> <Self::Accumulator as CubeType>::ExpandType

fn __expand_zero_accumulator( context: &mut CubeContext, acc: <Self::Accumulator as CubeType>::ExpandType, config: Self::Config, ) -> <() as CubeType>::ExpandType

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.