cubecl_linalg::matmul::components

Trait MatmulConfigFactory

pub trait MatmulConfigFactory:
    Send
    + Sync
    + 'static {
    type Config: MatmulConfig;
    type Input;

    // Required methods
    fn check_config(config: &Self::Config) -> Result<(), InvalidConfigError>;
    fn check_availability<R: Runtime, MP: MatmulPrecision>(
        _client: &ComputeClient<R::Server, R::Channel>,
        _config: &Self::Config,
    ) -> Result<(), MatmulAvailabilityError>;
    fn make_config(
        input: Self::Input,
        problem: &MatmulProblem,
        cube_dim: &CubeDim,
        cube_count: &CubeCount,
        advanced_config: &AdvancedConfig,
    ) -> Self::Config;
}

Provides configuration for a matmul kernel at any level

Required Associated Types

type Config: MatmulConfig

Configuration tailored to the matmul implementation

type Input

Required Methods

fn check_config(config: &Self::Config) -> Result<(), InvalidConfigError>

Asserts that the configuration for this matmul will lead to a valid computation

fn check_availability<R: Runtime, MP: MatmulPrecision>( _client: &ComputeClient<R::Server, R::Channel>, _config: &Self::Config, ) -> Result<(), MatmulAvailabilityError>

Checks if the client can handle the features used in this computation

fn make_config( input: Self::Input, problem: &MatmulProblem, cube_dim: &CubeDim, cube_count: &CubeCount, advanced_config: &AdvancedConfig, ) -> Self::Config

Create config for this matmul, given launch information

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.