cubecl_linalg::matmul::components

Trait StageDim

pub trait StageDim:
    'static
    + Send
    + Sync {
    // Required methods
    fn buffer_num_elements(&self) -> u32;
    fn num_tiles_x_dim(&self) -> u32;
    fn num_tiles_y_dim(&self) -> u32;
    fn tile_size_x_dim(&self) -> u32;
    fn tile_size_y_dim(&self) -> u32;

    // Provided methods
    fn total_elements(&self) -> u32 { ... }
    fn tile_num_elements(&self) -> u32 { ... }
    fn num_elements_x_dim(&self) -> u32 { ... }
    fn num_elements_y_dim(&self) -> u32 { ... }
    fn num_tiles(&self) -> u32 { ... }
}

Required Methods

fn buffer_num_elements(&self) -> u32

Number of elements in a buffer

fn num_tiles_x_dim(&self) -> u32

Returns the number of tiles across x dimension (rows)

fn num_tiles_y_dim(&self) -> u32

Returns the number of tiles across y dimension (cols)

fn tile_size_x_dim(&self) -> u32

Returns the dimension of a tile across x dimension (rows)

fn tile_size_y_dim(&self) -> u32

Returns the dimension of a tile across y dimension (col)

Provided Methods

fn total_elements(&self) -> u32

Returns the total number of elements of the stage

fn tile_num_elements(&self) -> u32

Returns the number of elements within one tile

fn num_elements_x_dim(&self) -> u32

Returns the number of elements across the x dimension

fn num_elements_y_dim(&self) -> u32

Returns the number of elements across the y dimension

fn num_tiles(&self) -> u32