cubecl_linalg

Module tensor

Contains basic tensor helpers.

Modules

index_offset_with_layout

Structs

Read

The read tag for virtual tensor.

ReadWrite

The read write tag for virtual tensor.

TensorHandle

Tensor representation containing a server handle as well as basic tensor metadata.,

VirtualTensor

Tensor representation that is decoupled from how the tensor is stored.

VirtualTensorExpand

Expand type for VirtualTensor.

Enums

MatrixLayout

Layout for matrix tensors, i.e. tensors whose interpretation is a bunch of batched matrices of 2 dimensions

Traits

VirtualTensorOperations

Trait to be implemented by a type that can become a virtual tensor.

VirtualTensorOperationsExpand

Expand trait for VirtualTensorOperations.

Functions

index_offset_with_layout

Returns the offset of the tensor corresponding to the layout tensor.

into_contiguous

Make a jit tensor contiguous.

into_contiguous_prefetch

Make a jit tensor contiguous.

matrix_layout

Return the layout of a matrix given the strides.