cubecl_linalg::tensor

Trait VirtualTensorOperationsExpand

pub trait VirtualTensorOperationsExpand<E: Numeric> {
    // Required methods
    fn __expand_read_method(
        &self,
        context: &mut CubeContext,
        index: ExpandElementTyped<u32>,
    ) -> ExpandElementTyped<Line<E>>;
    fn __expand_write_method(
        &self,
        context: &mut CubeContext,
        index: ExpandElementTyped<u32>,
        value: ExpandElementTyped<Line<E>>,
    );
    fn __expand_shape_method(
        &self,
        context: &mut CubeContext,
        axis: ExpandElementTyped<u32>,
    ) -> ExpandElementTyped<u32>;
    fn __expand_stride_method(
        &self,
        context: &mut CubeContext,
        axis: ExpandElementTyped<u32>,
    ) -> ExpandElementTyped<u32>;
    fn __expand_rank_method(
        &self,
        context: &mut CubeContext,
    ) -> ExpandElementTyped<u32>;
}

Expand trait for VirtualTensorOperations.

For now this needs to be manually implemented for any type that needs to become a virtual tensor.

Required Methods

fn __expand_read_method( &self, context: &mut CubeContext, index: ExpandElementTyped<u32>, ) -> ExpandElementTyped<Line<E>>

fn __expand_write_method( &self, context: &mut CubeContext, index: ExpandElementTyped<u32>, value: ExpandElementTyped<Line<E>>, )

fn __expand_shape_method( &self, context: &mut CubeContext, axis: ExpandElementTyped<u32>, ) -> ExpandElementTyped<u32>

fn __expand_stride_method( &self, context: &mut CubeContext, axis: ExpandElementTyped<u32>, ) -> ExpandElementTyped<u32>

fn __expand_rank_method( &self, context: &mut CubeContext, ) -> ExpandElementTyped<u32>